pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "GSIP"
version = "0.1.0"
description = "A scalable, multi-modal deep learning pipeline for robust and efficient inference on gigapixel-scale Earth Observation (EO) data."
authors = [{name = "Rati Vardiashvili", email = "rvardiashvili@constructor.university"}]
readme = "README.md"
license = {file = "LICENSE"}
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
requires-python = ">=3.9, <3.14"
dependencies = [
    "numpy",
    "torch",
    "rasterio",
    "tqdm",
    "lightning",
    "huggingface_hub",
    "matplotlib",
    "Pillow",
    "scipy",
    "configilm",
    "hydra-core",
    "omegaconf",
    "transformers",
    "lmdb",
    "psutil",
    "pandas",
    "safetensors",
    "timm",
    "nvidia-ml-py",
    "PyGObject",
]

[project.urls]
"Homepage" = "https://github.com/rvardiashvili/GeoTiff-Scalable-Analysis-Pipeline"

[project.scripts]
gsip = "src.cli:main"


[tool.setuptools.packages.find]
where = ["."]

[tool.setuptools.package-data]
"*" = ["*.yaml", "*.json", "*.png", "*.jpg"]
